разбирает код на модули
